[#32] set: :file_folder: router.js 분리
